+++ lib/src/request/get_accounts.g.dart
@@ -18,15 +18,3 @@ Map<String, dynamic> _$$_GetAccountsResultToJson(
     <String, dynamic>{
       'accounts': instance.accounts,
     };
-
-_$_AppAccount _$$_AppAccountFromJson(Map<String, dynamic> json) =>
-    _$_AppAccount(
-      name: json['name'] as String,
-      genesisAddress: json['genesisAddress'] as String,
-    );
-
-Map<String, dynamic> _$$_AppAccountToJson(_$_AppAccount instance) =>
-    <String, dynamic>{
-      'name': instance.name,
-      'genesisAddress': instance.genesisAddress,
-    };