+++ lib/src/request/get_current_account.g.dart
@@ -1,0 +1,21 @@
+// GENERATED CODE - DO NOT MODIFY BY HAND
+
+part of 'get_current_account.dart';
+
+// **************************************************************************
+// JsonSerializableGenerator
+// **************************************************************************
+
+_$_GetCurrentAccountResult _$$_GetCurrentAccountResultFromJson(
+        Map<String, dynamic> json) =>
+    _$_GetCurrentAccountResult(
+      name: json['name'] as String,
+      genesisAddress: json['genesisAddress'] as String,
+    );
+
+Map<String, dynamic> _$$_GetCurrentAccountResultToJson(
+        _$_GetCurrentAccountResult instance) =>
+    <String, dynamic>{
+      'name': instance.name,
+      'genesisAddress': instance.genesisAddress,
+    };